services/lucius-orchestrator/requirements.txt
fastapi>=0.110.0
uvicorn>=0.23.0
azure-data-tables>=12.4.0
azure-servicebus>=7.11.0
temporalio>=1.6.0
httpx>=0.24.0
